requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.1
django-admin-rangefilter==0.12.0
django-environ==0.11.2
django-import-export==3.3.5
et-xmlfile==1.1.0
gunicorn==21.2.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.9.0
tzdata==2023.4
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
